Introduce piddiplatsch.core package; move BaseProcessor, BaseRecord, HostingNode and shared model utilities; add shims in processing/models/records; tests+lint green
