+++ src/piddiplatsch/core/__init__.py
@@ -1,12 +1,16 @@
-from piddiplatsch.core.models import (
+from .models import (
     ALLOWED_CHECKSUM_METHODs,
     HostingNode,
     get_max_parts,
     strict_mode,
 )
+from .processing import BaseProcessor
+from .records import BaseRecord
 
 __all__ = [
     "ALLOWED_CHECKSUM_METHODs",
+    "BaseProcessor",
+    "BaseRecord",
     "HostingNode",
     "get_max_parts",
     "strict_mode",
